% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateByProteinChange.R
\name{annotateByProteinChange}
\alias{annotateByProteinChange}
\title{Pull annotations for mutations by protein change}
\usage{
annotateByProteinChange(
  api,
  mutations,
  by = c("entrezGeneId", "hugoSymbol"),
  getOncoTreeCODES = TRUE
)
}
\description{
Pull annotations for mutations by protein change
}
\examples{
library(cBioPortalData)
cbio <- cBioPortal()
mutations <- getDataByGenes(
    cbio, studyId = "acc_tcga", genes = 1:3, by = "entrezGeneId",
    molecularProfileIds = "acc_tcga_mutations",
    sampleListId = "acc_tcga_all"
)[["acc_tcga_mutations"]]

## Annotate the protein change
ok <- oncoKB(token = Sys.getenv("ONCOKB_TOKEN"))
annotateByProteinChange(
    ok, mutations, by = "entrezGeneId", getOncoTreeCODES = TRUE
)

}
