% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoKB.R
\name{cancerGeneList}
\alias{cancerGeneList}
\title{Obtain the OncoKB cancer gene list}
\usage{
cancerGeneList(api)
}
\arguments{
\item{api}{An OncoKB API instance as returned by \code{oncoKB()}}
}
\value{
A long tibble of genes with additional columns
}
\description{
Obtain the OncoKB cancer gene list
}
\examples{

oncokb <- oncoKB()
cancerGeneList(oncokb)

}
