% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoKB.R
\name{levelsOfEvidence}
\alias{levelsOfEvidence}
\title{Get the Levels of Evidence Tables from OncoKB}
\usage{
levelsOfEvidence(api)
}
\arguments{
\item{api}{An OncoKB API instance as returned by \code{oncoKB()}}
}
\value{
A \code{DataFrame} with metadata
}
\description{
The levels of evidence table is an \code{S4Vectors} \code{DataFrame} that includes
metadata consisting of app, api, data, and oncoTree version tags which were
used to generate the table.
}
