% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoKB.R
\name{oncoKB}
\alias{oncoKB}
\title{The R interface to the oncoKB API Data Service}
\usage{
oncoKB(
  hostname = "www.oncokb.org",
  protocol = "https",
  api. = .DEMO_ENDPOINT,
  token = character()
)
}
\arguments{
\item{hostname}{character(1) The internet location of the service
(default: 'www.cbioportal.org')}

\item{protocol}{character(1) The internet protocol used to access the
hostname (default: 'https')}

\item{api.}{character(1) The directory location of the API protocol within
the hostname (default: '/api/api-docs')}

\item{token}{character(1) The Authorization Bearer token e.g.,
"63eba81c-2591-4e15-9d1c-fb6e8e51e35d" or a path to text file.}
}
\value{
oncoKB: An API object of class 'oncoKB'
}
\description{
This section of the documentation lists the functions that
allow users to access the oncoKB API. The main representation of the
API can be obtained from the \code{oncoKB} function. The supporting
functions listed here give access to specific parts of the API and
allow the user to explore the API with individual calls.
}
\examples{
oncokb <- oncoKB()

## Authorization: Bearer token as a file
if (interactive())
  oncoKB(token = "~/onco_token.txt")

names(operations(oncokb))

}
